Move reusable components to new Dashboard folder
