--- frontend/src/pages/VulnerabilityScanDash/InfoLabel.tsx
@@ -1,48 +1,0 @@
-import React from 'react';
-import {
-  Box,
-  Typography,
-  Link as MuiLink,
-  TypographyProps
-} from '@mui/material';
-import { Link as RouterLink } from 'react-router-dom';
-import InfoTooltipIcon from './InfoTooltipIcon';
-
-type InfoLabelProps = {
-  label: string;
-  viewDetails?: boolean;
-  link?: string;
-  typographyVariant?: TypographyProps['variant'];
-  headingLevel?: 'h2' | 'h3' | 'p';
-};
-
-const InfoLabel: React.FC<InfoLabelProps> = ({
-  label,
-  viewDetails,
-  link,
-  typographyVariant = 'h2',
-  headingLevel = 'h2'
-}) => (
-  <Box display="flex" alignItems="center" justifyContent="space-between" p={0}>
-    <Box display="flex" alignItems="center">
-      <Typography
-        variant={typographyVariant}
-        component={headingLevel}
-        color="primary.darker"
-      >
-        {label}
-      </Typography>
-      <InfoTooltipIcon label={label} />
-    </Box>
-
-    {viewDetails && link && (
-      <MuiLink to={link} component={RouterLink}>
-        <Typography variant="link" component="p" fontWeight="bold">
-          View Details
-        </Typography>
-      </MuiLink>
-    )}
-  </Box>
-);
-
-export default InfoLabel;
