+++ frontend/src/pages/VulnerabilityScanDash/TopVulnerabilities.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom';
 import { Box, Link, Stack, Tooltip, Typography } from '@mui/material';
 import { Circle } from '@mui/icons-material';
 import { CVEItem } from 'types/vuln-scan-stats';
-import RoundedTable from 'components/RoundedTable';
+import RoundedTable from 'components/Dashboard/RoundedTable';
 import { severityColor } from 'utils/severityLevelColorMap';
 
 type ColumnConfig<T> = {
