+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -9,18 +9,19 @@ import {
 } from '@mui/material';
 import KeyMetrics from './KeyMetrics';
 import ScanningSummary from './ScanningSummary';
-import InfoLabel from './InfoLabel';
+import InfoLabel from 'components/Dashboard/InfoLabel';
+import GraphChip from 'components/Dashboard/GraphChip';
 import {
   // StatsTrendsRawData,
   vulnScanDataTransformed
 } from 'types/vuln-scan-stats';
 import { useAuthContext } from 'context';
-import GraphChip from './GraphChip';
 import { transformVulnScanData } from 'utils/transformVulnScanData';
 import RiskyOpenServices from './RiskyOpenServices';
 import TopVulnerableHosts from './TopVulnerableHosts';
 import TopVulnerabilities from './TopVulnerabilities';
 import SeverityByProminence from './SeverityByProminence';
+import infoIconContent from './infoIconContent.json';
 
 export interface VulnSeverities {
   label: string;
@@ -46,12 +47,12 @@ const InitialVSData = {
 };
 
 const formatDate = (date: Date): string => date.toISOString().split('T')[0];
-
 const today = formatDate(new Date());
-
 const yesterday = formatDate(
   new Date(new Date().setDate(new Date().getDate() - 1))
 );
+
+const tooltipContentJson = infoIconContent.infoIconContent;
 
 export const transformScanData = (rawData: any) => {
   return rawData.map((item: any) => ({
@@ -206,6 +207,7 @@ const VulnerabilityScan: React.FC = () => {
             label="Vulnerability Scanning Dashboard"
             typographyVariant="h1"
             viewDetails
+            tooltipContentJson={tooltipContentJson}
           />
         </Grid>
       </Grid>
@@ -263,7 +265,10 @@ const VulnerabilityScan: React.FC = () => {
             justifyContent="center"
           >
             <Grid size={{ xs: 12 }}>
-              <InfoLabel label="Latest Scanning Summary" />
+              <InfoLabel
+                label="Latest Scanning Summary"
+                tooltipContentJson={tooltipContentJson}
+              />
             </Grid>
             <Grid size={{ md: 12, lg: 5.5 }}>
               <ScanningSummary data={vulnScanData.vulnScanSummary} />
@@ -295,6 +300,7 @@ const VulnerabilityScan: React.FC = () => {
                 label="Known Exploited and Other Detected Vulnerabilities"
                 viewDetails
                 link="/inventory/vulnerabilities"
+                tooltipContentJson={tooltipContentJson}
               />
             </Grid>
           </Grid>
@@ -319,6 +325,7 @@ const VulnerabilityScan: React.FC = () => {
                       label="Severity by Prominence"
                       typographyVariant="h3"
                       headingLevel="h3"
+                      tooltipContentJson={tooltipContentJson}
                     />
                   </Grid>
                   <Grid size={{ xs: 12, lg: 5 }} justifyItems={'flex-end'}>
@@ -354,6 +361,7 @@ const VulnerabilityScan: React.FC = () => {
                       label="Top Vulnerabilities by Occurrence"
                       typographyVariant="h3"
                       headingLevel="h3"
+                      tooltipContentJson={tooltipContentJson}
                     />
                   </Grid>
                   <Grid size={{ xs: 12, lg: 3 }} justifyItems={'flex-end'}>
@@ -392,7 +400,10 @@ const VulnerabilityScan: React.FC = () => {
                 alignItems="center"
               >
                 <Grid size={{ xs: 12 }}>
-                  <InfoLabel label="Detected Services" />
+                  <InfoLabel
+                    label="Detected Services"
+                    tooltipContentJson={tooltipContentJson}
+                  />
                 </Grid>
               </Grid>
               <Grid container direction="column" spacing={2}>
@@ -411,6 +422,7 @@ const VulnerabilityScan: React.FC = () => {
                       label="Potentially Risky Services"
                       typographyVariant="h3"
                       headingLevel="h3"
+                      tooltipContentJson={tooltipContentJson}
                     />
                     <RiskyOpenServices data={vulnScanData.riskyServices} />
                   </Box>
@@ -438,6 +450,7 @@ const VulnerabilityScan: React.FC = () => {
                     label="Detected Hosts"
                     viewDetails
                     link="/inventory/domains"
+                    tooltipContentJson={tooltipContentJson}
                   />
                 </Grid>
               </Grid>
@@ -457,6 +470,7 @@ const VulnerabilityScan: React.FC = () => {
                       label="Top Vulnerable Hosts"
                       typographyVariant="h3"
                       headingLevel="h3"
+                      tooltipContentJson={tooltipContentJson}
                     />
                     <Grid container direction="row" paddingBottom={2}>
                       <GraphChip
